Automate IconName values
IconName values are now generated automatically based on the
webcomponents-icons folder from the node_modules. We are directly
reading from the modules installed in the demo sub-project.

We also added some gate checks in the github actions to be sure that the
icon values are in sync with the installed version.